Add admin guard clauses

+++ app/controllers/vehicles_controller.rb
@@ -3,12 +3,18 @@ class VehiclesController < ApplicationController
 
   # POST '/vehicles'
   def create
+    # Break out of this method if user isn't an admin
+    return unless @current_user.admin
+
     vehicle = Vehicle.create!(vehicle_params)
     render json: vehicle, status: :created
   end
 
   # DESTROY '/vehicles/[:id]
   def destroy
+    # Break out of this method if user isn't an admin
+    return unless @current_user.admin
+
     vehicle = find_vehicle
     vehicle.destroy
     head :no_content
@@ -16,11 +22,17 @@ class VehiclesController < ApplicationController
 
   # GET '/vehicles'
   def index
+    # Break out of this method if user isn't an admin
+    return unless @current_user.admin
+
     render json: Vehicle.all
   end
 
   # PATCH '/vehicles/[:id]
   def update
+    # Break out of this method if user isn't an admin
+    return unless @current_user.admin
+
     vehicle = find_vehicle
     if vehicle
       vehicle.update(vehicle_params)
